react.js no baka
Solves warning: React does not recognize the `requireAuth` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `requireauth` instead. If you accidentally passed it from a parent component, remove it from the DOM element. Error Component Stack