Revert "feat: refactor image loading to use object URLs; improve caching and implement revocation"
This reverts commit 93c8d0427c000200430e1e3fb031474761ec020c.

Reverted coz while it saves memory, in case we have way too many images,
we use CPU to generate URLs for each of them, which results in slowdown of animations.

In contrast, base64 strings are saved in cache, so a single cache access is all that is needed
to load these. The difference is as follows
- createObjectURL runs everytime after cache load
- base64 is already in cache so no more conversions needed

This minor difference seems to slow down the app for large no. of images, even though
base64 results in 33% more buffer sizes (should technically be slower)